+++ wmflib/tests/sphinx_checker.py
@@ -16,7 +16,7 @@ import wmflib
 DOC_API_BASE_PATH = 'doc/source/api'
 DOC_API_INDEX_PATH = os.path.join(DOC_API_BASE_PATH, 'index.rst')
 API_INDEX_PREFIX = '    wmflib.'
-EXCLUDED_NAMES = ('log')
+EXCLUDED_NAMES = ()
 
 
 def main(base_path):