+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 wmflib Changelog
 ----------------
 
+`v0.0.4`_ (2020-11-02)
+^^^^^^^^^^^^^^^^^^^^^^
+
+New features
+""""""""""""
+
+* requests: add new requests module that exposes an ``http_session()`` function that instantiate a requests's
+  ``Session`` with configurable default timeout, retry logic on some failures as well as setting a well formatted
+  User-Agent.
 
 `v0.0.3`_ (2020-10-23)
 ^^^^^^^^^^^^^^^^^^^^^^
@@ -33,3 +42,4 @@ New features
 .. _`v0.0.1`: https://github.com/wikimedia/operations-software-pywmflib/releases/tag/v0.0.1
 .. _`v0.0.2`: https://github.com/wikimedia/operations-software-pywmflib/releases/tag/v0.0.2
 .. _`v0.0.3`: https://github.com/wikimedia/operations-software-pywmflib/releases/tag/v0.0.3
+.. _`v0.0.4`: https://github.com/wikimedia/operations-software-pywmflib/releases/tag/v0.0.4